wip on loading scenes
a reasonably parsable wbml-ish scene is written out. pressing the load key 1 will clear the existing world, but we still need to implement the part where everything gets rehydrated from the WBML. extra credit after that will be to be able to load the scene additively, so that i can spawn multiple players easily. that'll be the fun id-rewriting part.